ai_backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
python-multipart
google-genai
Pillow